scx_lavd: Enable SCX_OPS_ENQ_MIGRATION_DISABLED.
By default, a migration-disabled task is put to the current CPU's local DSQ
with a _default_ time slice. Putting it into the local DSQ is fine but
assigning a default time slice is not because it could be too long (20 ms).

Hence, enable the SCX_OPS_ENQ_MIGRATION_DISABLED flag so migration-disabled
tasks also go through ops.enqueue(), and their time slices are assigned
explicitly.

Signed-off-by: Changwoo Min <changwoo@igalia.com>