scx_lavd: Boost the time slice of long-running tasks whenever possible.
When the system is not heavily loaded, so it can serve all tasks within the
targeted latency (slice_max_ns <= sys_stat.slice), we can consider boosting
the task's time slice if necessary. Additionally, if there are pinned tasks
waiting to run on this CPU, we do not boost the task's time slice to avoid
delaying the pinned task that cannot be run on another CPU.

If the task's avg_runtime is greater than slice_max_ns
(i.e., taskc->avg_runtime > slice_max_ns), that means the task could be
scheduled out due to a shorter time slice than required.

In this case, let's set the task's time slice to its avg_runtime to reduce
unnecessary involuntary context switching.

Signed-off-by: Changwoo Min <changwoo@igalia.com>