scx_lavd: Bounded relaxation of deadline enforcement (--relax-deadline).
Previously, scx_lavd always chose a task with the earliest deadline within
a domain by putting tasks to the per-domain DSQ and pulling them in a deadline
order. This design is good in terms of work conservation, but it has two
drawbacks; First, it is hard to control where a task actually runs. That is
because the task will be dispatched and executed on a CPU that finishes the
execution of the current task. That ruins the L1/L2 cache locality. More
importantly, since all CPUs in a domain are constantly harnessing the shared
per-domain DSQ, it makes the shared DSQ a contending hotspot, causing
fluctuations in the DSQ access time.

This commit partially addresses the above-mentioned problem by relaxing the
deadline enforcement within a certain bound. Instead of putting tasks to the
domain DSQ all the time, scx_lavd enqueues tasks to the chosen CPU's local DSQ
directly. This direct dispatch to local DSQ can effectively improve L1/L2 cache
locality and reduce the contention on the shared domain DSQ. However, the local
DSQ is FIFO-ordered, so a less critical task with a later deadline can run
earlier -- i.e., the deadline inversion can happen -- that will ruin tail
latency and make performance unpredictable.

To tame the deadline inversion problem, scx_lavd maintains the relaxation in
a bounded manner. It allows only N tasks to be queued into the local DSQ,
limiting the deadline violation upto the time slice of N tasks. When there are
more tasks in the local DSQ than the allowed limit (N), it enqueues the task to
the domain DSQ. To avoid the tasks in the domain DSQ being stalled,  it does
not put tasks to the local DSQ if there are tasks wanting to run on this CPU in
the domain DSQ (cpuc->nr_enq_non_local > 0). So scx_lavd consumes the locally
enqueued tasks first, then consumes the tasks on the domain DSQ next. When
there is no task wanting to run on this CPU in the domain DSQ
(cpuc->nr_enq_non_local == 0), it resumes the relaxed mode.

The slice boost and preemption mechanisms are updated to comply with the
relaxation. The slice boost will be canceled when there are pending tasks on
the local DSQ. This prevents locally enqueued tasks from waiting for execution
too long. In addition, the preemption mechanism is updated. It is meaningless
to kick a CPU other than where a task is locally enqueued. So the preemption
considers only the chosen CPU as a victim when a task is enqueued to the local
DSQ.

To specify N, we added an option, "--relaxed-deadline". The default value of
N is 2.

`schbench` shows decent improvement near saturation (say, 95% CPU utilization).

`schbench -L -m 2 -M auto -t 64 -n 0`

|                 | N = 0    | N = 2    |
|-----------------|----------|----------|
| average rps     | 568794.9 | 647113.3 |
| 99.0th latency  |      179 |      186 |
| 99.9th latency  |      204 |      215 |

`schbench -L -m 8 -M auto -t 64 -n 0`

|                 | N = 0     | N = 2     |
|-----------------|-----------|-----------|
| average rps     | 430927.87 | 897134.40 |
| 99.0th latency  |       417 |       279 |
| 99.9th latency  |       508 |       436 |

Signed-off-by: Changwoo Min <changwoo@igalia.com>