lib: cgroup_bw: Implement scx_cgroup_bw_throttled().
We test if a cgroup is throttled or not in a bottom-up manner: first at
the LLC level, then at the cgroup level, and finally at the subroot cgroup
level. Before asking the budget to the subroot cgroup level, update this
cgroup's runtime_total_sloppy to avoid spending the budget over its upper
bound (nquota_ub). We traverse the cgroup hierarchy in a post-order manner
(left, right, then root) and check the cgroup's level to efficiently sum
the runtime_total of all its descendants.

Signed-off-by: Changwoo Min <changwoo@igalia.com>