scx_lavd: Implement ops.deque() for throttling cancelling.
A throttled task may be dequeued by the sched_ext core for various reasons,
including sched_setaffinity(2), task migration, and so on. Eventually, such
a task will be enqueued again, breaking the internal red-black tree state of
the ATQ where it was. To avoid such a case, we should remove the dequeuing
task from the ATQ if it is in.

Signed-off-by: Emil Tsalapatis <etsal@meta.com>
Signed-off-by: Changwoo Min <changwoo@igalia.com>