scx_lavd: Revise the load comparison when deciding a sticky CPU/domain.
Previously, we compared the number of queued tasks when deciding on
a sticky CPU/domain. This approach checks only per-domain DSQ,
so it has a limitation that per-CPU DSQs are not adequately accounted for.
In addition, the queue length does not accurately reflect the load,
especially when CPU capacities are heterogeneous or when the queue length
is zero (i.e., when the system is underutilized).

We address this problem by comparing the scaled load of a domain ('sc_load'),
which accounts for each CPU's capacity and CPU utilization.

Signed-off-by: Changwoo Min <changwoo@igalia.com>